+++ tests/ts-loader.mjs
@@ -10,8 +10,13 @@ export async function resolve(specifier, context, defaultResolve) {
       return { url, shortCircuit: true };
     }
     if (specifier.startsWith('.') && !specifier.endsWith('.js') && !specifier.endsWith('.ts') && !specifier.endsWith('.tsx')) {
-      const url = new URL(specifier + '.ts', context.parentURL).href;
-      return { url, shortCircuit: true };
+      let url = new URL(specifier + '.ts', context.parentURL);
+      try {
+        await readFile(url);
+      } catch {
+        url = new URL(specifier + '.tsx', context.parentURL);
+      }
+      return { url: url.href, shortCircuit: true };
     }
     throw err;
   }
